// src/app/admin/(refine)/refine/interface.ts
export interface User {
  id: string;
  uid: string;
  name: string;
  email: string;
  phone: string;
  university: string;
  sizeJersey: string;
  foodLimitation: string;
  invitationCode: string;
  age: string;
  chronicDisease: string;
  drugAllergy: string;
  status: string;
  graduatedYear: string;
  faculty: string;
  imageUrl: string;
  lastEntered: string;
  registeredAt: string;
  role: string;
  education: string;
}